@@ -1,13 +1,13 @@
 module github.com/0xERR0R/blocky
 
-go 1.19
+go 1.20
 
 require (
-	github.com/abice/go-enum v0.5.4
-	github.com/alicebob/miniredis/v2 v2.30.0
+	github.com/abice/go-enum v0.5.5
+	github.com/alicebob/miniredis/v2 v2.30.1
 	github.com/asaskevich/EventBus v0.0.0-20200907212545-49d423059eef
-	github.com/avast/retry-go/v4 v4.3.2
-	github.com/creasty/defaults v1.6.0
+	github.com/avast/retry-go/v4 v4.3.3
+	github.com/creasty/defaults v1.7.0
 	github.com/go-chi/chi/v5 v5.0.8
 	github.com/go-chi/cors v1.2.1
 	github.com/go-redis/redis/v8 v8.11.5
