@@ -9,28 +9,28 @@ var _ = Describe("Caches", func() {
 	Describe("String StringCache", func() {
 		When("string StringCache was created", func() {
 			factory := newStringCacheFactory()
-			factory.AddEntry("google.com")
-			factory.AddEntry("apple.com")
-			factory.AddEntry("")
-			factory.AddEntry("google.com")
-			factory.AddEntry("APPLe.com")
+			factory.addEntry("google.com")
+			factory.addEntry("apple.com")
+			factory.addEntry("")
+			factory.addEntry("google.com")
+			factory.addEntry("APPLe.com")
 
-			cache := factory.Create()
+			cache := factory.create()
 
 			It("should match if StringCache contains exact string", func() {
-				Expect(cache.Contains("apple.com")).Should(BeTrue())
-				Expect(cache.Contains("google.com")).Should(BeTrue())
-				Expect(cache.Contains("www.google.com")).Should(BeFalse())
-				Expect(cache.Contains("")).Should(BeFalse())
+				Expect(cache.contains("apple.com")).Should(BeTrue())
+				Expect(cache.contains("google.com")).Should(BeTrue())
+				Expect(cache.contains("www.google.com")).Should(BeFalse())
+				Expect(cache.contains("")).Should(BeFalse())
 			})
 			It("should match case-insensitive", func() {
-				Expect(cache.Contains("aPPle.com")).Should(BeTrue())
-				Expect(cache.Contains("google.COM")).Should(BeTrue())
-				Expect(cache.Contains("www.google.com")).Should(BeFalse())
-				Expect(cache.Contains("")).Should(BeFalse())
+				Expect(cache.contains("aPPle.com")).Should(BeTrue())
+				Expect(cache.contains("google.COM")).Should(BeTrue())
+				Expect(cache.contains("www.google.com")).Should(BeFalse())
+				Expect(cache.contains("")).Should(BeFalse())
 			})
 			It("should return correct element count", func() {
-				Expect(cache.ElementCount()).Should(Equal(2))
+				Expect(cache.elementCount()).Should(Equal(2))
 			})
 		})
 	})
@@ -38,52 +38,30 @@ var _ = Describe("Caches", func() {
 	Describe("Regex StringCache", func() {
 		When("regex StringCache was created", func() {
 			factory := newRegexCacheFactory()
-			factory.AddEntry(".*google.com")
-			factory.AddEntry("^apple\\.(de|com)$")
-			factory.AddEntry("amazon")
+			factory.addEntry("/.*google.com/")
+			factory.addEntry("/^apple\\.(de|com)$/")
+			factory.addEntry("/amazon/")
 			// this is not a regex, will be ignored
-			factory.AddEntry("(wrongRegex")
-			cache := factory.Create()
+			factory.addEntry("/(wrongRegex/")
+			factory.addEntry("plaintext")
+			cache := factory.create()
 			It("should match if one regex in StringCache matches string", func() {
-				Expect(cache.Contains("google.com")).Should(BeTrue())
-				Expect(cache.Contains("google.coma")).Should(BeTrue())
-				Expect(cache.Contains("agoogle.com")).Should(BeTrue())
-				Expect(cache.Contains("www.google.com")).Should(BeTrue())
-				Expect(cache.Contains("apple.com")).Should(BeTrue())
-				Expect(cache.Contains("apple.de")).Should(BeTrue())
-				Expect(cache.Contains("apple.it")).Should(BeFalse())
-				Expect(cache.Contains("www.apple.com")).Should(BeFalse())
-				Expect(cache.Contains("applecom")).Should(BeFalse())
-				Expect(cache.Contains("www.amazon.com")).Should(BeTrue())
-				Expect(cache.Contains("amazon.com")).Should(BeTrue())
-				Expect(cache.Contains("myamazon.com")).Should(BeTrue())
+				Expect(cache.contains("google.com")).Should(BeTrue())
+				Expect(cache.contains("google.coma")).Should(BeTrue())
+				Expect(cache.contains("agoogle.com")).Should(BeTrue())
+				Expect(cache.contains("www.google.com")).Should(BeTrue())
+				Expect(cache.contains("apple.com")).Should(BeTrue())
+				Expect(cache.contains("apple.de")).Should(BeTrue())
+				Expect(cache.contains("apple.it")).Should(BeFalse())
+				Expect(cache.contains("www.apple.com")).Should(BeFalse())
+				Expect(cache.contains("applecom")).Should(BeFalse())
+				Expect(cache.contains("www.amazon.com")).Should(BeTrue())
+				Expect(cache.contains("amazon.com")).Should(BeTrue())
+				Expect(cache.contains("myamazon.com")).Should(BeTrue())
 			})
 			It("should return correct element count", func() {
-				Expect(cache.ElementCount()).Should(Equal(3))
-			})
-		})
-	})
-
-	Describe("Chained StringCache", func() {
-		When("chained StringCache was created", func() {
-			factory := NewChainedCacheFactory()
-			factory.AddEntry("/.*google.com/")
-			factory.AddEntry("/^apple\\.(de|com)$/")
-			factory.AddEntry("amazon.com")
-			cache := factory.Create()
-			It("should match if one regex in StringCache matches string", func() {
-				Expect(cache.Contains("google.com")).Should(BeTrue())
-				Expect(cache.Contains("google.coma")).Should(BeTrue())
-				Expect(cache.Contains("agoogle.com")).Should(BeTrue())
-				Expect(cache.Contains("www.google.com")).Should(BeTrue())
-				Expect(cache.Contains("apple.com")).Should(BeTrue())
-				Expect(cache.Contains("amazon.com")).Should(BeTrue())
-				Expect(cache.Contains("apple.de")).Should(BeTrue())
-				Expect(cache.Contains("www.apple.com")).Should(BeFalse())
-				Expect(cache.Contains("applecom")).Should(BeFalse())
-			})
-			It("should return correct element count", func() {
-				Expect(cache.ElementCount()).Should(Equal(3))
+				Expect(factory.count()).Should(Equal(3))
+				Expect(cache.elementCount()).Should(Equal(3))
 			})
 		})
 	})
