@@ -1,4 +1,4 @@
-//go:generate go run github.com/abice/go-enum -f=$GOFILE --marshal --names
+//go:generate go run github.com/abice/go-enum -f=$GOFILE --marshal --names --values
 package config
 
 import (
@@ -7,16 +7,12 @@ import (
 	"net"
 	"os"
 	"path/filepath"
-	"regexp"
-	"sort"
 	"strconv"
 	"strings"
 	"sync"
-	"time"
 
 	"github.com/miekg/dns"
-
-	"github.com/hako/durafmt"
+	"github.com/sirupsen/logrus"
 
 	"github.com/0xERR0R/blocky/log"
 	"github.com/creasty/defaults"
@@ -28,6 +24,16 @@ const (
 	tlsPort   = 853
 	httpsPort = 443
 )
+
+type Configurable interface {
+	// IsEnabled returns true when the receiver is configured.
+	IsEnabled() bool
+
+	// LogConfig logs the receiver's configuration.
+	//
+	// Calling this method when `IsEnabled` returns false is undefined.
+	LogConfig(*logrus.Entry)
+}
 
 // NetProtocol resolver protocol ENUM(
 // tcp+udp // TCP and UDP protocols
@@ -90,44 +96,6 @@ type StartStrategyType uint16
 // ENUM(clientIP,clientName,responseReason,responseAnswer,question,duration)
 type QueryLogField string
 
-type QType dns.Type
-
-func (c QType) String() string {
-	return dns.Type(c).String()
-}
-
-type QTypeSet map[QType]struct{}
-
-func NewQTypeSet(qTypes ...dns.Type) QTypeSet {
-	s := make(QTypeSet, len(qTypes))
-
-	for _, qType := range qTypes {
-		s.Insert(qType)
-	}
-
-	return s
-}
-
-func (s QTypeSet) Contains(qType dns.Type) bool {
-	_, found := s[QType(qType)]
-
-	return found
-}
-
-func (s *QTypeSet) Insert(qType dns.Type) {
-	if *s == nil {
-		*s = make(QTypeSet, 1)
-	}
-
-	(*s)[QType(qType)] = struct{}{}
-}
-
-type Duration time.Duration
-
-func (c Duration) String() string {
-	return durafmt.Parse(time.Duration(c)).String()
-}
-
 //nolint:gochecknoglobals
 var netDefaultPort = map[NetProtocol]uint16{
 	NetProtocolTcpUdp: udpPort,
@@ -135,82 +103,12 @@ var netDefaultPort = map[NetProtocol]uint16{
 	NetProtocolHttps:  httpsPort,
 }
 
-// Upstream is the definition of external DNS server
-type Upstream struct {
-	Net        NetProtocol
-	Host       string
-	Port       uint16
-	Path       string
-	CommonName string // Common Name to use for certificate verification; optional. "" uses .Host
-}
-
-// IsDefault returns true if u is the default value
-func (u *Upstream) IsDefault() bool {
-	return *u == Upstream{}
-}
-
-// String returns the string representation of u
-func (u Upstream) String() string {
-	if u.IsDefault() {
-		return "no upstream"
-	}
-
-	var sb strings.Builder
-
-	sb.WriteString(u.Net.String())
-	sb.WriteRune(':')
-
-	if u.Net == NetProtocolHttps {
-		sb.WriteString("//")
-	}
-
-	isIPv6 := strings.ContainsRune(u.Host, ':')
-	if isIPv6 {
-		sb.WriteRune('[')
-		sb.WriteString(u.Host)
-		sb.WriteRune(']')
-	} else {
-		sb.WriteString(u.Host)
-	}
-
-	if u.Port != netDefaultPort[u.Net] {
-		sb.WriteRune(':')
-		sb.WriteString(fmt.Sprint(u.Port))
-	}
-
-	if u.Path != "" {
-		sb.WriteString(u.Path)
-	}
-
-	return sb.String()
-}
-
-// UnmarshalYAML creates Upstream from YAML
-func (u *Upstream) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var s string
-	if err := unmarshal(&s); err != nil {
-		return err
-	}
-
-	upstream, err := ParseUpstream(s)
-	if err != nil {
-		return fmt.Errorf("can't convert upstream '%s': %w", s, err)
-	}
-
-	*u = upstream
-
-	return nil
-}
-
 // ListenConfig is a list of address(es) to listen on
 type ListenConfig []string
 
-// UnmarshalYAML creates ListenConfig from YAML
-func (l *ListenConfig) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var addresses string
-	if err := unmarshal(&addresses); err != nil {
-		return err
-	}
+// UnmarshalText implements `encoding.TextUnmarshaler`.
+func (l *ListenConfig) UnmarshalText(data []byte) error {
+	addresses := string(data)
 
 	*l = strings.Split(addresses, ",")
 
@@ -256,225 +154,11 @@ func (b *BootstrappedUpstreamConfig) UnmarshalYAML(unmarshal func(interface{}) e
 	return nil
 }
 
-// UnmarshalYAML creates ConditionalUpstreamMapping from YAML
-func (c *ConditionalUpstreamMapping) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var input map[string]string
-	if err := unmarshal(&input); err != nil {
-		return err
-	}
-
-	result := make(map[string][]Upstream, len(input))
-
-	for k, v := range input {
-		var upstreams []Upstream
-
-		for _, part := range strings.Split(v, ",") {
-			upstream, err := ParseUpstream(strings.TrimSpace(part))
-			if err != nil {
-				return fmt.Errorf("can't convert upstream '%s': %w", strings.TrimSpace(part), err)
-			}
-
-			upstreams = append(upstreams, upstream)
-		}
-
-		result[k] = upstreams
-	}
-
-	c.Upstreams = result
-
-	return nil
-}
-
-// UnmarshalYAML creates CustomDNSMapping from YAML
-func (c *CustomDNSMapping) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var input map[string]string
-	if err := unmarshal(&input); err != nil {
-		return err
-	}
-
-	result := make(map[string][]net.IP, len(input))
-
-	for k, v := range input {
-		var ips []net.IP
-
-		for _, part := range strings.Split(v, ",") {
-			ip := net.ParseIP(strings.TrimSpace(part))
-			if ip == nil {
-				return fmt.Errorf("invalid IP address '%s'", part)
-			}
-
-			ips = append(ips, ip)
-		}
-
-		result[k] = ips
-	}
-
-	c.HostIPs = result
-
-	return nil
-}
-
-// UnmarshalYAML creates Duration from YAML. If no unit is used, uses minutes
-func (c *Duration) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var input string
-	if err := unmarshal(&input); err != nil {
-		return err
-	}
-
-	if minutes, err := strconv.Atoi(input); err == nil {
-		// duration is defined as number without unit
-		// use minutes to ensure back compatibility
-		*c = Duration(time.Duration(minutes) * time.Minute)
-
-		return nil
-	}
-
-	duration, err := time.ParseDuration(input)
-	if err == nil {
-		*c = Duration(duration)
-
-		return nil
-	}
-
-	return err
-}
-
-func (c *QType) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var input string
-	if err := unmarshal(&input); err != nil {
-		return err
-	}
-
-	t, found := dns.StringToType[input]
-	if !found {
-		types := make([]string, 0, len(dns.StringToType))
-		for k := range dns.StringToType {
-			types = append(types, k)
-		}
-
-		sort.Strings(types)
-
-		return fmt.Errorf("unknown DNS query type: '%s'. Please use following types '%s'",
-			input, strings.Join(types, ", "))
-	}
-
-	*c = QType(t)
-
-	return nil
-}
-
-func (s *QTypeSet) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var input []QType
-	if err := unmarshal(&input); err != nil {
-		return err
-	}
-
-	*s = make(QTypeSet, len(input))
-
-	for _, qType := range input {
-		(*s)[qType] = struct{}{}
-	}
-
-	return nil
-}
-
-var validDomain = regexp.MustCompile(
-	`^(([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9\-]*[a-zA-Z0-9])\.)*([A-Za-z0-9]|[A-Za-z0-9][A-Za-z0-9\-]*[A-Za-z0-9])$`)
-
-// ParseUpstream creates new Upstream from passed string in format [net]:host[:port][/path][#commonname]
-func ParseUpstream(upstream string) (Upstream, error) {
-	var path string
-
-	var port uint16
-
-	commonName, upstream := extractCommonName(upstream)
-
-	n, upstream := extractNet(upstream)
-
-	path, upstream = extractPath(upstream)
-
-	host, portString, err := net.SplitHostPort(upstream)
-
-	// string contains host:port
-	if err == nil {
-		p, err := ConvertPort(portString)
-		if err != nil {
-			err = fmt.Errorf("can't convert port to number (1 - 65535) %w", err)
-
-			return Upstream{}, err
-		}
-
-		port = p
-	} else {
-		// only host, use default port
-		host = upstream
-		port = netDefaultPort[n]
-
-		// trim any IPv6 brackets
-		host = strings.TrimPrefix(host, "[")
-		host = strings.TrimSuffix(host, "]")
-	}
-
-	// validate hostname or ip
-	if ip := net.ParseIP(host); ip == nil {
-		// is not IP
-		if !validDomain.MatchString(host) {
-			return Upstream{}, fmt.Errorf("wrong host name '%s'", host)
-		}
-	}
-
-	return Upstream{
-		Net:        n,
-		Host:       host,
-		Port:       port,
-		Path:       path,
-		CommonName: commonName,
-	}, nil
-}
-
-func extractCommonName(in string) (string, string) {
-	upstream, cn, _ := strings.Cut(in, "#")
-
-	return cn, upstream
-}
-
-func extractPath(in string) (path, upstream string) {
-	slashIdx := strings.Index(in, "/")
-
-	if slashIdx >= 0 {
-		path = in[slashIdx:]
-		upstream = in[:slashIdx]
-	} else {
-		upstream = in
-	}
-
-	return
-}
-
-func extractNet(upstream string) (NetProtocol, string) {
-	tcpUDPPrefix := NetProtocolTcpUdp.String() + ":"
-	if strings.HasPrefix(upstream, tcpUDPPrefix) {
-		return NetProtocolTcpUdp, upstream[len(tcpUDPPrefix):]
-	}
-
-	tcpTLSPrefix := NetProtocolTcpTls.String() + ":"
-	if strings.HasPrefix(upstream, tcpTLSPrefix) {
-		return NetProtocolTcpTls, upstream[len(tcpTLSPrefix):]
-	}
-
-	httpsPrefix := NetProtocolHttps.String() + ":"
-	if strings.HasPrefix(upstream, httpsPrefix) {
-		return NetProtocolHttps, strings.TrimPrefix(upstream[len(httpsPrefix):], "//")
-	}
-
-	return NetProtocolTcpUdp, upstream
-}
-
 // Config main configuration
 //
 //nolint:maligned
 type Config struct {
-	Upstream            UpstreamConfig            `yaml:"upstream"`
+	Upstream            ParallelBestConfig        `yaml:"upstream"`
 	UpstreamTimeout     Duration                  `yaml:"upstreamTimeout" default:"2s"`
 	ConnectIPVersion    IPVersion                 `yaml:"connectIPVersion"`
 	CustomDNS           CustomDNSConfig           `yaml:"customDNS"`
@@ -483,7 +167,7 @@ type Config struct {
 	ClientLookup        ClientLookupConfig        `yaml:"clientLookup"`
 	Caching             CachingConfig             `yaml:"caching"`
 	QueryLog            QueryLogConfig            `yaml:"queryLog"`
-	Prometheus          PrometheusConfig          `yaml:"prometheus"`
+	Prometheus          MetricsConfig             `yaml:"prometheus"`
 	Redis               RedisConfig               `yaml:"redis"`
 	Log                 log.Config                `yaml:"log"`
 	Ports               PortsConfig               `yaml:"ports"`
@@ -494,7 +178,7 @@ type Config struct {
 	KeyFile             string                    `yaml:"keyFile"`
 	BootstrapDNS        BootstrapDNSConfig        `yaml:"bootstrapDns"`
 	HostsFile           HostsFileConfig           `yaml:"hostsFile"`
-	FqdnOnly            bool                      `yaml:"fqdnOnly" default:"false"`
+	FqdnOnly            FqdnOnlyConfig            `yaml:",inline"`
 	Filtering           FilteringConfig           `yaml:"filtering"`
 	Ede                 EdeConfig                 `yaml:"ede"`
 	// Deprecated
@@ -508,7 +192,7 @@ type Config struct {
 	// Deprecated
 	LogTimestamp bool `yaml:"logTimestamp" default:"true"`
 	// Deprecated
-	DNSPorts ListenConfig `yaml:"port" default:"[\"53\"]"`
+	DNSPorts ListenConfig `yaml:"port" default:"53"`
 	// Deprecated
 	HTTPPorts ListenConfig `yaml:"httpPort"`
 	// Deprecated
@@ -518,10 +202,17 @@ type Config struct {
 }
 
 type PortsConfig struct {
-	DNS   ListenConfig `yaml:"dns" default:"[\"53\"]"`
+	DNS   ListenConfig `yaml:"dns" default:"53"`
 	HTTP  ListenConfig `yaml:"http"`
 	HTTPS ListenConfig `yaml:"https"`
 	TLS   ListenConfig `yaml:"tls"`
+}
+
+func (c *PortsConfig) LogConfig(logger *logrus.Entry) {
+	logger.Infof("DNS   = %s", c.DNS)
+	logger.Infof("TLS   = %s", c.TLS)
+	logger.Infof("HTTP  = %s", c.HTTP)
+	logger.Infof("HTTPS = %s", c.HTTPS)
 }
 
 // split in two types to avoid infinite recursion. See `BootstrapDNSConfig.UnmarshalYAML`.
@@ -539,105 +230,6 @@ type (
 	}
 )
 
-// PrometheusConfig contains the config values for prometheus
-type PrometheusConfig struct {
-	Enable bool   `yaml:"enable" default:"false"`
-	Path   string `yaml:"path" default:"/metrics"`
-}
-
-// UpstreamConfig upstream server configuration
-type UpstreamConfig struct {
-	ExternalResolvers map[string][]Upstream `yaml:",inline"`
-}
-
-// RewriteConfig custom DNS configuration
-type RewriteConfig struct {
-	Rewrite          map[string]string `yaml:"rewrite"`
-	FallbackUpstream bool              `yaml:"fallbackUpstream" default:"false"`
-}
-
-// CustomDNSConfig custom DNS configuration
-type CustomDNSConfig struct {
-	RewriteConfig       `yaml:",inline"`
-	CustomTTL           Duration         `yaml:"customTTL" default:"1h"`
-	Mapping             CustomDNSMapping `yaml:"mapping"`
-	FilterUnmappedTypes bool             `yaml:"filterUnmappedTypes" default:"true"`
-}
-
-// CustomDNSMapping mapping for the custom DNS configuration
-type CustomDNSMapping struct {
-	HostIPs map[string][]net.IP
-}
-
-// ConditionalUpstreamConfig conditional upstream configuration
-type ConditionalUpstreamConfig struct {
-	RewriteConfig `yaml:",inline"`
-	Mapping       ConditionalUpstreamMapping `yaml:"mapping"`
-}
-
-// ConditionalUpstreamMapping mapping for conditional configuration
-type ConditionalUpstreamMapping struct {
-	Upstreams map[string][]Upstream
-}
-
-// BlockingConfig configuration for query blocking
-type BlockingConfig struct {
-	BlackLists        map[string][]string `yaml:"blackLists"`
-	WhiteLists        map[string][]string `yaml:"whiteLists"`
-	ClientGroupsBlock map[string][]string `yaml:"clientGroupsBlock"`
-	BlockType         string              `yaml:"blockType" default:"ZEROIP"`
-	BlockTTL          Duration            `yaml:"blockTTL" default:"6h"`
-	DownloadTimeout   Duration            `yaml:"downloadTimeout" default:"60s"`
-	DownloadAttempts  uint                `yaml:"downloadAttempts" default:"3"`
-	DownloadCooldown  Duration            `yaml:"downloadCooldown" default:"1s"`
-	RefreshPeriod     Duration            `yaml:"refreshPeriod" default:"4h"`
-	// Deprecated
-	FailStartOnListError  bool              `yaml:"failStartOnListError" default:"false"`
-	ProcessingConcurrency uint              `yaml:"processingConcurrency" default:"4"`
-	StartStrategy         StartStrategyType `yaml:"startStrategy" default:"blocking"`
-}
-
-// ClientLookupConfig configuration for the client lookup
-type ClientLookupConfig struct {
-	ClientnameIPMapping map[string][]net.IP `yaml:"clients"`
-	Upstream            Upstream            `yaml:"upstream"`
-	SingleNameOrder     []uint              `yaml:"singleNameOrder"`
-}
-
-// CachingConfig configuration for domain caching
-type CachingConfig struct {
-	MinCachingTime        Duration `yaml:"minTime"`
-	MaxCachingTime        Duration `yaml:"maxTime"`
-	CacheTimeNegative     Duration `yaml:"cacheTimeNegative" default:"30m"`
-	MaxItemsCount         int      `yaml:"maxItemsCount"`
-	Prefetching           bool     `yaml:"prefetching"`
-	PrefetchExpires       Duration `yaml:"prefetchExpires" default:"2h"`
-	PrefetchThreshold     int      `yaml:"prefetchThreshold" default:"5"`
-	PrefetchMaxItemsCount int      `yaml:"prefetchMaxItemsCount"`
-}
-
-func (c *CachingConfig) EnablePrefetch() {
-	const day = 24 * time.Hour
-
-	if c.MaxCachingTime == 0 {
-		// make sure resolver gets enabled
-		c.MaxCachingTime = Duration(day)
-	}
-
-	c.Prefetching = true
-	c.PrefetchThreshold = 0
-}
-
-// QueryLogConfig configuration for the query logging
-type QueryLogConfig struct {
-	Target           string          `yaml:"target"`
-	Type             QueryLogType    `yaml:"type"`
-	LogRetentionDays uint64          `yaml:"logRetentionDays"`
-	CreationAttempts int             `yaml:"creationAttempts" default:"3"`
-	CreationCooldown Duration        `yaml:"creationCooldown" default:"2s"`
-	Fields           []QueryLogField `yaml:"fields"`
-}
-
 // RedisConfig configuration for the redis connection
 type RedisConfig struct {
 	Address            string   `yaml:"address"`
@@ -652,19 +244,23 @@ type RedisConfig struct {
 	SentinelAddresses  []string `yaml:"sentinelAddresses"`
 }
 
-type HostsFileConfig struct {
-	Filepath       string   `yaml:"filePath"`
-	HostsTTL       Duration `yaml:"hostsTTL" default:"1h"`
-	RefreshPeriod  Duration `yaml:"refreshPeriod" default:"1h"`
-	FilterLoopback bool     `yaml:"filterLoopback"`
-}
+type (
+	FqdnOnlyConfig = toEnable
+	EdeConfig      = toEnable
+)
 
-type FilteringConfig struct {
-	QueryTypes QTypeSet `yaml:"queryTypes"`
-}
-
-type EdeConfig struct {
+type toEnable struct {
 	Enable bool `yaml:"enable" default:"false"`
+}
+
+// IsEnabled implements `config.Configurable`.
+func (c *toEnable) IsEnabled() bool {
+	return c.Enable
+}
+
+// LogConfig implements `config.Configurable`.
+func (c *toEnable) LogConfig(logger *logrus.Entry) {
+	logger.Info("enabled")
 }
 
 //nolint:gochecknoglobals
