@@ -51,7 +51,7 @@ var _ = Describe("External lists and query blocking", func() {
 								HaveTTL(BeNumerically("==", 123)),
 							))
 
-					Expect(getContainerLogs(blocky)).Should(ContainElement(ContainSubstring("error during file processing")))
+					Expect(getContainerLogs(blocky)).Should(ContainElement(ContainSubstring("cannot open source: ")))
 				})
 			})
 			Context("startStrategy = failOnError", func() {
@@ -77,7 +77,7 @@ var _ = Describe("External lists and query blocking", func() {
 				})
 
 				It("should fail to start", func() {
-					Eventually(blocky.IsRunning, "30s").Should(BeFalse())
+					Eventually(blocky.IsRunning, "5s", "2ms").Should(BeFalse())
 
 					Expect(getContainerLogs(blocky)).
 						Should(ContainElement(ContainSubstring("Error: can't start server: 1 error occurred")))
