@@ -72,7 +72,7 @@ var _ = Describe("Redis configuration tests", func() {
 			It("2nd instance of blocky should use cache from redis", func() {
 				msg := util.NewMsgWithQuestion("google.de.", A)
 				By("Query first blocky instance, should store cache in redis", func() {
-					Expect(doDNSRequest(blocky1, msg)).
+					Eventually(doDNSRequest, "5s", "2ms").WithArguments(blocky1, msg).
 						Should(
 							SatisfyAll(
 								BeDNSRecord("google.de.", A, "1.2.3.4"),
@@ -81,7 +81,7 @@ var _ = Describe("Redis configuration tests", func() {
 				})
 
 				By("Check redis, must contain one cache entry", func() {
-					Eventually(dbSize).WithArguments(redisClient).Should(BeNumerically("==", 1))
+					Eventually(dbSize, "5s", "2ms").WithArguments(redisClient).Should(BeNumerically("==", 1))
 				})
 
 				By("Shutdown the upstream DNS server", func() {
@@ -89,7 +89,7 @@ var _ = Describe("Redis configuration tests", func() {
 				})
 
 				By("Query second blocky instance, should use cache from redis", func() {
-					Expect(doDNSRequest(blocky2, msg)).
+					Eventually(doDNSRequest, "5s", "2ms").WithArguments(blocky2, msg).
 						Should(
 							SatisfyAll(
 								BeDNSRecord("google.de.", A, "1.2.3.4"),
@@ -124,7 +124,7 @@ var _ = Describe("Redis configuration tests", func() {
 			It("should load cache from redis after start", func() {
 				msg := util.NewMsgWithQuestion("google.de.", A)
 				By("Query first blocky instance, should store cache in redis\"", func() {
-					Expect(doDNSRequest(blocky1, msg)).
+					Eventually(doDNSRequest, "5s", "2ms").WithArguments(blocky1, msg).
 						Should(
 							SatisfyAll(
 								BeDNSRecord("google.de.", A, "1.2.3.4"),
@@ -156,7 +156,7 @@ var _ = Describe("Redis configuration tests", func() {
 				})
 
 				By("Query second blocky instance", func() {
-					Expect(doDNSRequest(blocky2, msg)).
+					Eventually(doDNSRequest, "5s", "2ms").WithArguments(blocky2, msg).
 						Should(
 							SatisfyAll(
 								BeDNSRecord("google.de.", A, "1.2.3.4"),
