@@ -35,7 +35,7 @@ func createBlockHandler(cfg config.BlockingConfig) (blockHandler, error) {
 		return nxDomainBlockHandler{}, nil
 	}
 
-	blockTime := uint32(time.Duration(cfg.BlockTTL).Seconds())
+	blockTime := cfg.BlockTTL.SecondsU32()
 
 	if strings.EqualFold(cfgBlockType, "ZEROIP") {
 		return zeroIPBlockHandler{
@@ -78,17 +78,18 @@ type status struct {
 
 // BlockingResolver checks request's question (domain name) against black and white lists
 type BlockingResolver struct {
+	configurable[*config.BlockingConfig]
 	NextResolver
+	typed
+
 	blacklistMatcher    *lists.ListCache
 	whitelistMatcher    *lists.ListCache
-	cfg                 config.BlockingConfig
 	blockHandler        blockHandler
 	whitelistOnlyGroups map[string]bool
 	status              *status
 	clientGroupsBlock   map[string][]string
 	redisClient         *redis.Client
-	redisEnabled        bool
-	fqdnIPCache         expirationcache.ExpiringCache
+	fqdnIPCache         expirationcache.ExpiringCache[[]net.IP]
 }
 
 // NewBlockingResolver returns a new configured instance of the resolver
@@ -100,7 +101,7 @@ func NewBlockingResolver(
 		return nil, err
 	}
 
-	refreshPeriod := time.Duration(cfg.RefreshPeriod)
+	refreshPeriod := cfg.RefreshPeriod.ToDuration()
 	downloader := createDownloader(cfg, bootstrap)
 	blacklistMatcher, blErr := lists.NewListCache(lists.ListCacheTypeBlacklist, cfg.BlackLists,
 		refreshPeriod, downloader, cfg.ProcessingConcurrency,
@@ -118,7 +119,7 @@ func NewBlockingResolver(
 	cgb := make(map[string][]string, len(cfg.ClientGroupsBlock))
 
 	for identifier, cfgGroups := range cfg.ClientGroupsBlock {
-		for _, ipart := range strings.Split(identifier, ",") {
+		for _, ipart := range strings.Split(strings.ToLower(identifier), ",") {
 			existingGroups, found := cgb[ipart]
 			if found {
 				cgb[ipart] = append(existingGroups, cfgGroups...)
@@ -129,8 +130,10 @@ func NewBlockingResolver(
 	}
 
 	res := &BlockingResolver{
+		configurable: withConfig(&cfg),
+		typed:        withType("blocking"),
+
 		blockHandler:        blockHandler,
-		cfg:                 cfg,
 		blacklistMatcher:    blacklistMatcher,
 		whitelistMatcher:    whitelistMatcher,
 		whitelistOnlyGroups: whitelistOnlyGroups,
@@ -140,10 +143,9 @@ func NewBlockingResolver(
 		},
 		clientGroupsBlock: cgb,
 		redisClient:       redis,
-		redisEnabled:      (redis != nil),
 	}
 
-	if res.redisEnabled {
+	if res.redisClient != nil {
 		setupRedisEnabledSubscriber(res)
 	}
 
@@ -156,27 +158,25 @@ func NewBlockingResolver(
 
 func createDownloader(cfg config.BlockingConfig, bootstrap *Bootstrap) *lists.HTTPDownloader {
 	return lists.NewDownloader(
-		lists.WithTimeout(time.Duration(cfg.DownloadTimeout)),
+		lists.WithTimeout(cfg.DownloadTimeout.ToDuration()),
 		lists.WithAttempts(cfg.DownloadAttempts),
-		lists.WithCooldown(time.Duration(cfg.DownloadCooldown)),
+		lists.WithCooldown(cfg.DownloadCooldown.ToDuration()),
 		lists.WithTransport(bootstrap.NewHTTPTransport()),
 	)
 }
 
 func setupRedisEnabledSubscriber(c *BlockingResolver) {
-	logger := log.PrefixedLog("blocking_resolver")
-
 	go func() {
 		for em := range c.redisClient.EnabledChannel {
 			if em != nil {
-				logger.Debug("Received state from redis: ", em)
+				c.log().Debug("Received state from redis: ", em)
 
 				if em.State {
 					c.internalEnableBlocking()
 				} else {
 					err := c.internalDisableBlocking(em.Duration, em.Groups)
 					if err != nil {
-						logger.Warn("Blocking couldn't be disabled:", err)
+						c.log().Warn("Blocking couldn't be disabled:", err)
 					}
 				}
 			}
@@ -213,7 +213,7 @@ func (r *BlockingResolver) retrieveAllBlockingGroups() []string {
 func (r *BlockingResolver) EnableBlocking() {
 	r.internalEnableBlocking()
 
-	if r.redisEnabled {
+	if r.redisClient != nil {
 		r.redisClient.PublishEnabled(&redis.EnabledMessage{State: true})
 	}
 }
@@ -232,7 +232,7 @@ func (r *BlockingResolver) internalEnableBlocking() {
 // DisableBlocking deactivates the blocking for a particular duration (or forever if 0).
 func (r *BlockingResolver) DisableBlocking(duration time.Duration, disableGroups []string) error {
 	err := r.internalDisableBlocking(duration, disableGroups)
-	if err == nil && r.redisEnabled {
+	if err == nil && r.redisClient != nil {
 		r.redisClient.PublishEnabled(&redis.EnabledMessage{
 			State:    false,
 			Duration: duration,
@@ -329,39 +329,15 @@ func (r *BlockingResolver) handleBlocked(logger *logrus.Entry,
 	return &model.Response{Res: response, RType: model.ResponseTypeBLOCKED, Reason: reason}, nil
 }
 
-// Configuration returns the current resolver configuration
-func (r *BlockingResolver) Configuration() (result []string) {
-	if len(r.cfg.ClientGroupsBlock) == 0 {
-		return configDisabled
-	}
+// LogConfig implements `config.Configurable`.
+func (r *BlockingResolver) LogConfig(logger *logrus.Entry) {
+	r.cfg.LogConfig(logger)
 
-	result = append(result, "clientGroupsBlock")
-	for key, val := range r.cfg.ClientGroupsBlock {
-		result = append(result, fmt.Sprintf("  %s = \"%s\"", key, strings.Join(val, ";")))
-	}
+	logger.Info("blacklist cache entries:")
+	log.WithIndent(logger, "  ", r.blacklistMatcher.LogConfig)
 
-	blockType := r.cfg.BlockType
-	result = append(result, fmt.Sprintf("blockType = \"%s\"", blockType))
-
-	if blockType != "NXDOMAIN" {
-		result = append(result, fmt.Sprintf("blockTTL = %s", r.cfg.BlockTTL.String()))
-	}
-
-	result = append(result, fmt.Sprintf("downloadTimeout = %s", r.cfg.DownloadTimeout.String()))
-
-	result = append(result, fmt.Sprintf("FailStartOnListError = %t", r.cfg.FailStartOnListError))
-
-	result = append(result, "blacklist:")
-	for _, c := range r.blacklistMatcher.Configuration() {
-		result = append(result, fmt.Sprintf("  %s", c))
-	}
-
-	result = append(result, "whitelist:")
-	for _, c := range r.whitelistMatcher.Configuration() {
-		result = append(result, fmt.Sprintf("  %s", c))
-	}
-
-	return result
+	logger.Info("whitelist cache entries:")
+	log.WithIndent(logger, "  ", r.whitelistMatcher.LogConfig)
 }
 
 func (r *BlockingResolver) hasWhiteListOnlyAllowed(groupsToCheck []string) bool {
@@ -384,8 +360,8 @@ func (r *BlockingResolver) handleBlacklist(groupsToCheck []string,
 		domain := util.ExtractDomain(question)
 		logger := logger.WithField("domain", domain)
 
-		if whitelisted, group := r.matches(groupsToCheck, r.whitelistMatcher, domain); whitelisted {
-			logger.WithField("group", group).Debugf("domain is whitelisted")
+		if groups := r.matches(groupsToCheck, r.whitelistMatcher, domain); len(groups) > 0 {
+			logger.WithField("groups", groups).Debugf("domain is whitelisted")
 
 			resp, err := r.next.Resolve(request)
 
@@ -398,8 +374,8 @@ func (r *BlockingResolver) handleBlacklist(groupsToCheck []string,
 			return true, resp, err
 		}
 
-		if blocked, group := r.matches(groupsToCheck, r.blacklistMatcher, domain); blocked {
-			resp, err := r.handleBlocked(logger, request, question, fmt.Sprintf("BLOCKED (%s)", group))
+		if groups := r.matches(groupsToCheck, r.blacklistMatcher, domain); len(groups) > 0 {
+			resp, err := r.handleBlocked(logger, request, question, fmt.Sprintf("BLOCKED (%s)", strings.Join(groups, ",")))
 
 			return true, resp, err
 		}
@@ -428,10 +404,11 @@ func (r *BlockingResolver) Resolve(request *model.Request) (*model.Response, err
 			if len(entryToCheck) > 0 {
 				logger := logger.WithField("response_entry", entryToCheck)
 
-				if whitelisted, group := r.matches(groupsToCheck, r.whitelistMatcher, entryToCheck); whitelisted {
-					logger.WithField("group", group).Debugf("%s is whitelisted", tName)
-				} else if blocked, group := r.matches(groupsToCheck, r.blacklistMatcher, entryToCheck); blocked {
-					return r.handleBlocked(logger, request, request.Req.Question[0], fmt.Sprintf("BLOCKED %s (%s)", tName, group))
+				if groups := r.matches(groupsToCheck, r.whitelistMatcher, entryToCheck); len(groups) > 0 {
+					logger.WithField("groups", groups).Debugf("%s is whitelisted", tName)
+				} else if groups := r.matches(groupsToCheck, r.blacklistMatcher, entryToCheck); len(groups) > 0 {
+					return r.handleBlocked(logger, request, request.Req.Question[0], fmt.Sprintf("BLOCKED %s (%s)", tName,
+						strings.Join(groups, ",")))
 				}
 			}
 		}
@@ -496,10 +473,9 @@ func (r *BlockingResolver) groupsToCheckForClient(request *model.Request) []stri
 		if util.CidrContainsIP(clientIdentifier, request.ClientIP) {
 			groups = append(groups, groupsByCidr...)
 		} else if isFQDN(clientIdentifier) && r.fqdnIPCache != nil {
-			clIps, _ := r.fqdnIPCache.Get(clientIdentifier)
-			if clIps != nil {
-				ips := clIps.([]net.IP)
-				for _, ip := range ips {
+			ips, _ := r.fqdnIPCache.Get(clientIdentifier)
+			if ips != nil {
+				for _, ip := range *ips {
 					if ip.Equal(request.ClientIP) {
 						groups = append(groups, groupsByCidr...)
 					}
@@ -528,15 +504,12 @@ func (r *BlockingResolver) groupsToCheckForClient(request *model.Request) []stri
 
 func (r *BlockingResolver) matches(groupsToCheck []string, m lists.Matcher,
 	domain string,
-) (blocked bool, group string) {
+) (group []string) {
 	if len(groupsToCheck) > 0 {
-		found, group := m.Match(domain, groupsToCheck)
-		if found {
-			return true, group
-		}
+		return m.Match(domain, groupsToCheck)
 	}
 
-	return false, ""
+	return []string{}
 }
 
 type blockHandler interface {
@@ -593,8 +566,12 @@ func (b ipBlockHandler) handleBlock(question dns.Question, response *dns.Msg) {
 	}
 }
 
-func (r *BlockingResolver) queryForFQIdentifierIPs(identifier string) (result []net.IP, ttl time.Duration) {
-	prefixedLog := log.PrefixedLog("FQDNClientIdentifierCache")
+func (r *BlockingResolver) queryForFQIdentifierIPs(identifier string) (*[]net.IP, time.Duration) {
+	prefixedLog := log.WithPrefix(r.log(), "client_id_cache")
+
+	var result []net.IP
+
+	var ttl time.Duration
 
 	for _, qType := range []uint16{dns.TypeA, dns.TypeAAAA} {
 		resp, err := r.next.Resolve(&model.Request{
@@ -618,7 +595,7 @@ func (r *BlockingResolver) queryForFQIdentifierIPs(identifier string) (result []
 		}
 	}
 
-	return
+	return &result, ttl
 }
 
 func (r *BlockingResolver) initFQDNIPCache() {
@@ -631,8 +608,8 @@ func (r *BlockingResolver) initFQDNIPCache() {
 		identifiers = append(identifiers, identifier)
 	}
 
-	r.fqdnIPCache = expirationcache.NewCache(expirationcache.WithCleanUpInterval(defaultBlockingCleanUpInterval),
-		expirationcache.WithOnExpiredFn(func(key string) (val interface{}, ttl time.Duration) {
+	r.fqdnIPCache = expirationcache.NewCache(expirationcache.WithCleanUpInterval[[]net.IP](defaultBlockingCleanUpInterval),
+		expirationcache.WithOnExpiredFn(func(key string) (val *[]net.IP, ttl time.Duration) {
 			return r.queryForFQIdentifierIPs(key)
 		}))
 
