@@ -7,6 +7,7 @@ import (
 	"github.com/0xERR0R/blocky/config"
 	. "github.com/0xERR0R/blocky/evt"
 	. "github.com/0xERR0R/blocky/helpertest"
+	"github.com/0xERR0R/blocky/log"
 	. "github.com/0xERR0R/blocky/model"
 	"github.com/0xERR0R/blocky/redis"
 	"github.com/0xERR0R/blocky/util"
@@ -27,6 +28,12 @@ var _ = Describe("CachingResolver", func() {
 		mockAnswer *dns.Msg
 	)
 
+	Describe("Type", func() {
+		It("follows conventions", func() {
+			expectValidResolverType(sut)
+		})
+	})
+
 	BeforeEach(func() {
 		sutConfig = config.CachingConfig{}
 		if err := defaults.Set(&sutConfig); err != nil {
@@ -40,6 +47,22 @@ var _ = Describe("CachingResolver", func() {
 		m = &mockResolver{}
 		m.On("Resolve", mock.Anything).Return(&Response{Res: mockAnswer}, nil)
 		sut.Next(m)
+	})
+
+	Describe("IsEnabled", func() {
+		It("is false", func() {
+			Expect(sut.IsEnabled()).Should(BeFalse())
+		})
+	})
+
+	Describe("LogConfig", func() {
+		It("should log something", func() {
+			logger, hook := log.NewMockEntry()
+
+			sut.LogConfig(logger)
+
+			Expect(hook.Calls).ShouldNot(BeEmpty())
+		})
 	})
 
 	Describe("Caching responses", func() {
@@ -58,7 +81,7 @@ var _ = Describe("CachingResolver", func() {
 				prefetchThreshold := 5
 				configureCaches(sut, &sutConfig)
 				sut.resultCache = expirationcache.NewCache(
-					expirationcache.WithCleanUpInterval(100*time.Millisecond),
+					expirationcache.WithCleanUpInterval[cacheValue](100*time.Millisecond),
 					expirationcache.WithOnExpiredFn(sut.onExpired))
 
 				domainPrefetched := make(chan string, 1)
@@ -102,6 +125,15 @@ var _ = Describe("CachingResolver", func() {
 							BeDNSRecord("example.com.", A, "123.122.121.120"),
 							HaveTTL(BeNumerically("<=", 2))))
 				Eventually(prefetchHitDomain, "4s").Should(Receive(Equal("example.com")))
+			})
+			When("threshold is 0", func() {
+				BeforeEach(func() {
+					sutConfig.PrefetchThreshold = 0
+				})
+
+				It("should always prefetch", func() {
+					Expect(sut.shouldPrefetch("domain.tld")).Should(BeTrue())
+				})
 			})
 		})
 		When("min caching time is defined", func() {
@@ -364,6 +396,10 @@ var _ = Describe("CachingResolver", func() {
 				})
 
 				It("response should be cached", func() {
+					By("default config should enable negative caching", func() {
+						Expect(sutConfig.CacheTimeNegative).Should(BeNumerically(">", 0))
+					})
+
 					By("first request", func() {
 						Expect(sut.Resolve(newRequest("example.com.", AAAA))).
 							Should(SatisfyAll(
@@ -491,43 +527,6 @@ var _ = Describe("CachingResolver", func() {
 					// still one call to resolver
 					Expect(m.Calls).Should(HaveLen(1))
 				})
-			})
-		})
-	})
-
-	Describe("Configuration output", func() {
-		When("resolver is enabled", func() {
-			BeforeEach(func() {
-				sutConfig = config.CachingConfig{}
-			})
-			It("should return configuration", func() {
-				c := sut.Configuration()
-				Expect(len(c)).Should(BeNumerically(">", 1))
-			})
-		})
-
-		When("resolver is disabled", func() {
-			BeforeEach(func() {
-				sutConfig = config.CachingConfig{
-					MaxCachingTime: config.Duration(time.Minute * -1),
-				}
-			})
-			It("should return 'disabled'", func() {
-				c := sut.Configuration()
-				Expect(c).Should(ContainElement(configStatusDisabled))
-			})
-		})
-
-		When("prefetching is enabled", func() {
-			BeforeEach(func() {
-				sutConfig = config.CachingConfig{
-					Prefetching: true,
-				}
-			})
-			It("should return configuration", func() {
-				c := sut.Configuration()
-				Expect(len(c)).Should(BeNumerically(">", 1))
-				Expect(c).Should(ContainElement(ContainSubstring("prefetchThreshold")))
 			})
 		})
 	})
