@@ -1,7 +1,6 @@
 package resolver
 
 import (
-	"fmt"
 	"strings"
 
 	"github.com/0xERR0R/blocky/config"
@@ -15,7 +14,10 @@ import (
 
 // ConditionalUpstreamResolver delegates DNS question to other DNS resolver dependent on domain name in question
 type ConditionalUpstreamResolver struct {
+	configurable[*config.ConditionalUpstreamConfig]
 	NextResolver
+	typed
+
 	mapping map[string]Resolver
 }
 
@@ -26,10 +28,13 @@ func NewConditionalUpstreamResolver(
 	m := make(map[string]Resolver, len(cfg.Mapping.Upstreams))
 
 	for domain, upstream := range cfg.Mapping.Upstreams {
-		upstreams := make(map[string][]config.Upstream)
-		upstreams[upstreamDefaultCfgName] = upstream
+		pbCfg := config.ParallelBestConfig{
+			ExternalResolvers: config.ParallelBestMapping{
+				upstreamDefaultCfgName: upstream,
+			},
+		}
 
-		r, err := NewParallelBestResolver(upstreams, bootstrap, shouldVerifyUpstreams)
+		r, err := NewParallelBestResolver(pbCfg, bootstrap, shouldVerifyUpstreams)
 		if err != nil {
 			return nil, err
 		}
@@ -37,20 +42,14 @@ func NewConditionalUpstreamResolver(
 		m[strings.ToLower(domain)] = r
 	}
 
-	return &ConditionalUpstreamResolver{mapping: m}, nil
-}
+	r := ConditionalUpstreamResolver{
+		configurable: withConfig(&cfg),
+		typed:        withType("conditional_upstream"),
 
-// Configuration returns current configuration
-func (r *ConditionalUpstreamResolver) Configuration() (result []string) {
-	if len(r.mapping) == 0 {
-		return configDisabled
+		mapping: m,
 	}
 
-	for key, val := range r.mapping {
-		result = append(result, fmt.Sprintf("%s = \"%s\"", key, val))
-	}
-
-	return
+	return &r, nil
 }
 
 func (r *ConditionalUpstreamResolver) processRequest(request *model.Request) (bool, *model.Response, error) {
