@@ -7,77 +7,47 @@ import (
 )
 
 type EdeResolver struct {
+	configurable[*config.EdeConfig]
 	NextResolver
-	config config.EdeConfig
+	typed
 }
 
 func NewEdeResolver(cfg config.EdeConfig) ChainedResolver {
 	return &EdeResolver{
-		config: cfg,
+		configurable: withConfig(&cfg),
+		typed:        withType("extended_error_code"),
 	}
 }
 
 func (r *EdeResolver) Resolve(request *model.Request) (*model.Response, error) {
+	if !r.cfg.Enable {
+		return r.next.Resolve(request)
+	}
+
 	resp, err := r.next.Resolve(request)
 	if err != nil {
 		return nil, err
 	}
 
-	if r.config.Enable {
-		addExtraReasoning(resp)
-	}
+	r.addExtraReasoning(resp)
 
 	return resp, nil
 }
 
-func (r *EdeResolver) Configuration() (result []string) {
-	if !r.config.Enable {
-		return configDisabled
+func (r *EdeResolver) addExtraReasoning(res *model.Response) {
+	infocode := res.RType.ToExtendedErrorCode()
+
+	if infocode == dns.ExtendedErrorCodeOther {
+		// dns.ExtendedErrorCodeOther seams broken in some clients
+		return
 	}
 
-	return configEnabled
-}
-
-func addExtraReasoning(res *model.Response) {
-	// dns.ExtendedErrorCodeOther seams broken in some clients
-	infocode := convertToExtendedErrorCode(res.RType)
-	if infocode > 0 {
-		opt := new(dns.OPT)
-		opt.Hdr.Name = "."
-		opt.Hdr.Rrtype = dns.TypeOPT
-		opt.Option = append(opt.Option, convertExtendedError(res, infocode))
-		res.Res.Extra = append(res.Res.Extra, opt)
-	}
-}
-
-func convertExtendedError(input *model.Response, infocode uint16) *dns.EDNS0_EDE {
-	return &dns.EDNS0_EDE{
+	opt := new(dns.OPT)
+	opt.Hdr.Name = "."
+	opt.Hdr.Rrtype = dns.TypeOPT
+	opt.Option = append(opt.Option, &dns.EDNS0_EDE{
 		InfoCode:  infocode,
-		ExtraText: input.Reason,
-	}
-}
-
-func convertToExtendedErrorCode(input model.ResponseType) uint16 {
-	switch input {
-	case model.ResponseTypeRESOLVED:
-		return dns.ExtendedErrorCodeOther
-	case model.ResponseTypeCACHED:
-		return dns.ExtendedErrorCodeCachedError
-	case model.ResponseTypeCONDITIONAL:
-		return dns.ExtendedErrorCodeForgedAnswer
-	case model.ResponseTypeCUSTOMDNS:
-		return dns.ExtendedErrorCodeForgedAnswer
-	case model.ResponseTypeHOSTSFILE:
-		return dns.ExtendedErrorCodeForgedAnswer
-	case model.ResponseTypeNOTFQDN:
-		return dns.ExtendedErrorCodeBlocked
-	case model.ResponseTypeBLOCKED:
-		return dns.ExtendedErrorCodeBlocked
-	case model.ResponseTypeFILTERED:
-		return dns.ExtendedErrorCodeFiltered
-	case model.ResponseTypeSPECIAL:
-		return dns.ExtendedErrorCodeFiltered
-	default:
-		return dns.ExtendedErrorCodeOther
-	}
+		ExtraText: res.Reason,
+	})
+	res.Res.Extra = append(res.Res.Extra, opt)
 }
