@@ -1,10 +1,6 @@
 package resolver
 
 import (
-	"fmt"
-	"sort"
-	"strings"
-
 	"github.com/0xERR0R/blocky/config"
 	"github.com/0xERR0R/blocky/model"
 	"github.com/miekg/dns"
@@ -13,13 +9,21 @@ import (
 // FilteringResolver filters DNS queries (for example can drop all AAAA query)
 // returns empty ANSWER with NOERROR
 type FilteringResolver struct {
+	configurable[*config.FilteringConfig]
 	NextResolver
-	queryTypes config.QTypeSet
+	typed
+}
+
+func NewFilteringResolver(cfg config.FilteringConfig) ChainedResolver {
+	return &FilteringResolver{
+		configurable: withConfig(&cfg),
+		typed:        withType("filtering"),
+	}
 }
 
 func (r *FilteringResolver) Resolve(request *model.Request) (*model.Response, error) {
 	qType := request.Req.Question[0].Qtype
-	if r.queryTypes.Contains(dns.Type(qType)) {
+	if r.cfg.QueryTypes.Contains(dns.Type(qType)) {
 		response := new(dns.Msg)
 		response.SetRcode(request.Req, dns.RcodeSuccess)
 
@@ -27,28 +31,4 @@ func (r *FilteringResolver) Resolve(request *model.Request) (*model.Response, er
 	}
 
 	return r.next.Resolve(request)
-}
-
-func (r *FilteringResolver) Configuration() (result []string) {
-	if len(r.queryTypes) == 0 {
-		return configDisabled
-	}
-
-	qTypes := make([]string, 0, len(r.queryTypes))
-
-	for qType := range r.queryTypes {
-		qTypes = append(qTypes, qType.String())
-	}
-
-	sort.Strings(qTypes)
-
-	result = append(result, fmt.Sprintf("filtering query Types: '%v'", strings.Join(qTypes, ", ")))
-
-	return
-}
-
-func NewFilteringResolver(cfg config.FilteringConfig) ChainedResolver {
-	return &FilteringResolver{
-		queryTypes: cfg.QueryTypes,
-	}
 }
