@@ -1,63 +1,100 @@
 package resolver
 
 import (
+	"context"
 	"fmt"
 	"net"
 	"os"
-	"strings"
 	"time"
 
 	"github.com/0xERR0R/blocky/config"
-	"github.com/0xERR0R/blocky/log"
+	"github.com/0xERR0R/blocky/lists/parsers"
 	"github.com/0xERR0R/blocky/model"
 	"github.com/0xERR0R/blocky/util"
 	"github.com/miekg/dns"
 	"github.com/sirupsen/logrus"
 )
 
-//nolint:gochecknoglobals
-var (
-	_, loopback4, _ = net.ParseCIDR("127.0.0.0/8")
-	loopback6       = net.ParseIP("::1")
-)
-
 const (
-	hostsFileResolverLogger = "hosts_file_resolver"
+	// reduce initial capacity so we don't waste memory if there are less entries than before
+	memReleaseFactor = 2
 )
 
 type HostsFileResolver struct {
+	configurable[*config.HostsFileConfig]
 	NextResolver
-	HostsFilePath  string
-	hosts          []host
-	ttl            uint32
-	refreshPeriod  time.Duration
-	filterLoopback bool
+	typed
+
+	hosts splitHostsFileData
+}
+
+type HostsFileEntry = parsers.HostsFileEntry
+
+func NewHostsFileResolver(cfg config.HostsFileConfig) *HostsFileResolver {
+	r := HostsFileResolver{
+		configurable: withConfig(&cfg),
+		typed:        withType("hosts_file"),
+	}
+
+	if err := r.parseHostsFile(context.Background()); err != nil {
+		r.log().Errorf("disabling hosts file resolving due to error: %s", err)
+
+		r.cfg.Filepath = "" // don't try parsing the file again
+	} else {
+		go r.periodicUpdate()
+	}
+
+	return &r
+}
+
+// LogConfig implements `config.Configurable`.
+func (r *HostsFileResolver) LogConfig(logger *logrus.Entry) {
+	r.cfg.LogConfig(logger)
+
+	logger.Infof("cache entries = %d", r.hosts.len())
 }
 
 func (r *HostsFileResolver) handleReverseDNS(request *model.Request) *model.Response {
 	question := request.Req.Question[0]
-	if question.Qtype == dns.TypePTR {
-		response := new(dns.Msg)
-		response.SetReply(request.Req)
+	if question.Qtype != dns.TypePTR {
+		return nil
+	}
 
-		for _, host := range r.hosts {
-			raddr, _ := dns.ReverseAddr(host.IP.String())
+	questionIP, err := util.ParseIPFromArpaAddr(question.Name)
+	if err != nil {
+		// ignore the parse error, and pass the request down the chain
+		return nil
+	}
 
-			if raddr == question.Name {
-				ptr := new(dns.PTR)
-				ptr.Ptr = dns.Fqdn(host.Hostname)
-				ptr.Hdr = util.CreateHeader(question, r.ttl)
-				response.Answer = append(response.Answer, ptr)
+	if r.cfg.FilterLoopback && questionIP.IsLoopback() {
+		// skip the search: we won't find anything
+		return nil
+	}
 
-				for _, alias := range host.Aliases {
-					ptrAlias := new(dns.PTR)
-					ptrAlias.Ptr = dns.Fqdn(alias)
-					ptrAlias.Hdr = util.CreateHeader(question, r.ttl)
-					response.Answer = append(response.Answer, ptrAlias)
-				}
+	// search only in the hosts with an IP version that matches the question
+	hostsData := r.hosts.v4
+	if questionIP.To4() == nil {
+		hostsData = r.hosts.v6
+	}
 
-				return &model.Response{Res: response, RType: model.ResponseTypeHOSTSFILE, Reason: "HOSTS FILE"}
+	for host, hostData := range hostsData.hosts {
+		if hostData.IP.Equal(questionIP) {
+			response := new(dns.Msg)
+			response.SetReply(request.Req)
+
+			ptr := new(dns.PTR)
+			ptr.Ptr = dns.Fqdn(host)
+			ptr.Hdr = util.CreateHeader(question, r.cfg.HostsTTL.SecondsU32())
+			response.Answer = append(response.Answer, ptr)
+
+			for _, alias := range hostData.Aliases {
+				ptrAlias := new(dns.PTR)
+				ptrAlias.Ptr = dns.Fqdn(alias)
+				ptrAlias.Hdr = ptr.Hdr
+				response.Answer = append(response.Answer, ptrAlias)
 			}
+
+			return &model.Response{Res: response, RType: model.ResponseTypeHOSTSFILE, Reason: "HOSTS FILE"}
 		}
 	}
 
@@ -65,9 +102,7 @@ func (r *HostsFileResolver) handleReverseDNS(request *model.Request) *model.Resp
 }
 
 func (r *HostsFileResolver) Resolve(request *model.Request) (*model.Response, error) {
-	logger := log.WithPrefix(request.Log, hostsFileResolverLogger)
-
-	if r.HostsFilePath == "" {
+	if r.cfg.Filepath == "" {
 		return r.next.Resolve(request)
 	}
 
@@ -76,150 +111,77 @@ func (r *HostsFileResolver) Resolve(request *model.Request) (*model.Response, er
 		return reverseResp, nil
 	}
 
-	if len(r.hosts) != 0 {
-		response := new(dns.Msg)
-		response.SetReply(request.Req)
+	question := request.Req.Question[0]
+	domain := util.ExtractDomain(question)
 
-		question := request.Req.Question[0]
-		domain := util.ExtractDomain(question)
+	response := r.resolve(request.Req, question, domain)
+	if response != nil {
+		r.log().WithFields(logrus.Fields{
+			"answer": util.AnswerToString(response.Answer),
+			"domain": domain,
+		}).Debugf("returning hosts file entry")
 
-		for _, host := range r.hosts {
-			response.Answer = append(response.Answer, r.processHostEntry(host, domain, question)...)
-		}
-
-		if len(response.Answer) > 0 {
-			logger.WithFields(logrus.Fields{
-				"answer": util.AnswerToString(response.Answer),
-				"domain": domain,
-			}).Debugf("returning hosts file entry")
-
-			return &model.Response{Res: response, RType: model.ResponseTypeHOSTSFILE, Reason: "HOSTS FILE"}, nil
-		}
+		return &model.Response{Res: response, RType: model.ResponseTypeHOSTSFILE, Reason: "HOSTS FILE"}, nil
 	}
 
-	logger.WithField("resolver", Name(r.next)).Trace("go to next resolver")
+	r.log().WithField("resolver", Name(r.next)).Trace("go to next resolver")
 
 	return r.next.Resolve(request)
 }
 
-func (r *HostsFileResolver) processHostEntry(host host, domain string, question dns.Question) (result []dns.RR) {
-	if host.Hostname == domain {
-		if isSupportedType(host.IP, question) {
-			rr, _ := util.CreateAnswerFromQuestion(question, host.IP, r.ttl)
-			result = append(result, rr)
-		}
-	}
-
-	for _, alias := range host.Aliases {
-		if alias == domain {
-			if isSupportedType(host.IP, question) {
-				rr, _ := util.CreateAnswerFromQuestion(question, host.IP, r.ttl)
-				result = append(result, rr)
-			}
-		}
-	}
-
-	return
-}
-
-func (r *HostsFileResolver) Configuration() (result []string) {
-	if r.HostsFilePath == "" || len(r.hosts) == 0 {
-		return configDisabled
-	}
-
-	result = append(result, fmt.Sprintf("file path: %s", r.HostsFilePath))
-	result = append(result, fmt.Sprintf("TTL: %d", r.ttl))
-	result = append(result, fmt.Sprintf("refresh period: %s", r.refreshPeriod.String()))
-	result = append(result, fmt.Sprintf("filter loopback addresses: %t", r.filterLoopback))
-
-	return
-}
-
-func NewHostsFileResolver(cfg config.HostsFileConfig) *HostsFileResolver {
-	r := HostsFileResolver{
-		HostsFilePath:  cfg.Filepath,
-		ttl:            uint32(time.Duration(cfg.HostsTTL).Seconds()),
-		refreshPeriod:  time.Duration(cfg.RefreshPeriod),
-		filterLoopback: cfg.FilterLoopback,
-	}
-
-	if err := r.parseHostsFile(); err != nil {
-		logger := log.PrefixedLog(hostsFileResolverLogger)
-		logger.Warnf("cannot parse hosts file: %s, hosts file resolving is disabled", r.HostsFilePath)
-		r.HostsFilePath = ""
-	} else {
-		go r.periodicUpdate()
-	}
-
-	return &r
-}
-
-type host struct {
-	IP       net.IP
-	Hostname string
-	Aliases  []string
-}
-
-//nolint:funlen
-func (r *HostsFileResolver) parseHostsFile() error {
-	const minColumnCount = 2
-
-	if r.HostsFilePath == "" {
+func (r *HostsFileResolver) resolve(req *dns.Msg, question dns.Question, domain string) *dns.Msg {
+	ip := r.hosts.getIP(dns.Type(question.Qtype), domain)
+	if ip == nil {
 		return nil
 	}
 
-	buf, err := os.ReadFile(r.HostsFilePath)
+	rr, _ := util.CreateAnswerFromQuestion(question, ip, r.cfg.HostsTTL.SecondsU32())
+
+	response := new(dns.Msg)
+	response.SetReply(req)
+	response.Answer = []dns.RR{rr}
+
+	return response
+}
+
+func (r *HostsFileResolver) parseHostsFile(ctx context.Context) error {
+	const maxErrorsPerFile = 5
+
+	if r.cfg.Filepath == "" {
+		return nil
+	}
+
+	f, err := os.Open(r.cfg.Filepath)
 	if err != nil {
 		return err
 	}
+	defer f.Close()
 
-	newHosts := make([]host, 0)
+	newHosts := newSplitHostsDataWithSameCapacity(r.hosts)
 
-	for _, line := range strings.Split(string(buf), "\n") {
-		trimmed := strings.TrimSpace(line)
+	p := parsers.AllowErrors(parsers.HostsFile(f), maxErrorsPerFile)
+	p.OnErr(func(err error) {
+		r.log().Warnf("error parsing %s: %s, trying to continue", r.cfg.Filepath, err)
+	})
 
-		if len(trimmed) == 0 || trimmed[0] == '#' {
-			// Skip empty and commented lines
-			continue
+	err = parsers.ForEach[*HostsFileEntry](ctx, p, func(entry *HostsFileEntry) error {
+		if len(entry.Interface) != 0 {
+			// Ignore entries with a specific interface: we don't restrict what clients/interfaces we serve entries to,
+			// so this avoids returning entries that can't be accessed by the client.
+			return nil
 		}
 
-		// Find comment symbol at the end of the line
-		var fields []string
-
-		end := strings.IndexRune(trimmed, '#')
-
-		if end == -1 {
-			fields = strings.Fields(trimmed)
-		} else {
-			fields = strings.Fields(trimmed[:end])
+		// Ignore loopback, if so configured
+		if r.cfg.FilterLoopback && (entry.IP.IsLoopback() || entry.Name == "localhost") {
+			return nil
 		}
 
-		if len(fields) < minColumnCount {
-			// Skip invalid entry
-			continue
-		}
+		newHosts.add(entry)
 
-		if net.ParseIP(fields[0]) == nil {
-			// Skip invalid IP address
-			continue
-		}
-
-		var h host
-		h.IP = net.ParseIP(fields[0])
-		h.Hostname = fields[1]
-
-		// Check if loopback
-		if r.filterLoopback && (loopback4.Contains(h.IP) || loopback6.Equal(h.IP)) {
-			continue
-		}
-
-		if len(fields) > minColumnCount {
-			for i := 2; i < len(fields); i++ {
-				h.Aliases = append(h.Aliases, fields[i])
-			}
-		}
-
-		newHosts = append(newHosts, h)
+		return nil
+	})
+	if err != nil {
+		return fmt.Errorf("error parsing %s: %w", r.cfg.Filepath, err) // err is parsers.ErrTooManyErrors
 	}
 
 	r.hosts = newHosts
@@ -228,17 +190,101 @@ func (r *HostsFileResolver) parseHostsFile() error {
 }
 
 func (r *HostsFileResolver) periodicUpdate() {
-	if r.refreshPeriod > 0 {
-		ticker := time.NewTicker(r.refreshPeriod)
+	if r.cfg.RefreshPeriod.ToDuration() > 0 {
+		ticker := time.NewTicker(r.cfg.RefreshPeriod.ToDuration())
 		defer ticker.Stop()
 
 		for {
 			<-ticker.C
 
-			logger := log.PrefixedLog(hostsFileResolverLogger)
-			logger.WithField("file", r.HostsFilePath).Debug("refreshing hosts file")
+			r.log().WithField("file", r.cfg.Filepath).Debug("refreshing hosts file")
 
-			util.LogOnError("can't refresh hosts file: ", r.parseHostsFile())
+			util.LogOnError("can't refresh hosts file: ", r.parseHostsFile(context.Background()))
 		}
+	}
+}
+
+// stores hosts file data for IP versions separately
+//
+// Makes finding an IP for a question faster.
+// Especially reverse lookups where we have to iterate through
+// all the known hosts.
+type splitHostsFileData struct {
+	v4 hostsFileData
+	v6 hostsFileData
+}
+
+func newSplitHostsDataWithSameCapacity(other splitHostsFileData) splitHostsFileData {
+	return splitHostsFileData{
+		v4: newHostsDataWithSameCapacity(other.v4),
+		v6: newHostsDataWithSameCapacity(other.v6),
+	}
+}
+
+func (d splitHostsFileData) isEmpty() bool {
+	return d.len() == 0
+}
+
+func (d splitHostsFileData) len() int {
+	return d.v4.len() + d.v6.len()
+}
+
+func (d splitHostsFileData) getIP(qType dns.Type, domain string) net.IP {
+	switch uint16(qType) {
+	case dns.TypeA:
+		return d.v4.getIP(domain)
+	case dns.TypeAAAA:
+		return d.v6.getIP(domain)
+	}
+
+	return nil
+}
+
+func (d splitHostsFileData) add(entry *parsers.HostsFileEntry) {
+	if entry.IP.To4() != nil {
+		d.v4.add(entry)
+	} else {
+		d.v6.add(entry)
+	}
+}
+
+type hostsFileData struct {
+	hosts   map[string]hostData
+	aliases map[string]net.IP
+}
+
+type hostData struct {
+	IP      net.IP
+	Aliases []string
+}
+
+func newHostsDataWithSameCapacity(other hostsFileData) hostsFileData {
+	return hostsFileData{
+		hosts:   make(map[string]hostData, len(other.hosts)/memReleaseFactor),
+		aliases: make(map[string]net.IP, len(other.aliases)/memReleaseFactor),
+	}
+}
+
+func (d hostsFileData) len() int {
+	return len(d.hosts) + len(d.aliases)
+}
+
+func (d hostsFileData) getIP(hostname string) net.IP {
+	if hostData, ok := d.hosts[hostname]; ok {
+		return hostData.IP
+	}
+
+	if ip, ok := d.aliases[hostname]; ok {
+		return ip
+	}
+
+	return nil
+}
+
+func (d hostsFileData) add(entry *parsers.HostsFileEntry) {
+	d.hosts[entry.Name] = hostData{entry.IP, entry.Aliases}
+
+	for _, alias := range entry.Aliases {
+		d.aliases[alias] = entry.IP
 	}
 }
