@@ -2,12 +2,19 @@ package resolver
 
 import (
 	. "github.com/0xERR0R/blocky/helpertest"
+	"github.com/0xERR0R/blocky/log"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
 
 var _ = Describe("NoOpResolver", func() {
 	var sut NoOpResolver
+
+	Describe("Type", func() {
+		It("follows conventions", func() {
+			expectValidResolverType(sut)
+		})
+	})
 
 	BeforeEach(func() {
 		sut = NewNoOpResolver().(NoOpResolver)
@@ -21,10 +28,19 @@ var _ = Describe("NoOpResolver", func() {
 		})
 	})
 
-	Describe("Configuration output", func() {
-		It("returns nothing", func() {
-			c := sut.Configuration()
-			Expect(c).Should(BeNil())
+	Describe("IsEnabled", func() {
+		It("is true", func() {
+			Expect(sut.IsEnabled()).Should(BeTrue())
+		})
+	})
+
+	Describe("LogConfig", func() {
+		It("should not log anything", func() {
+			logger, hook := log.NewMockEntry()
+
+			sut.LogConfig(logger)
+
+			Expect(hook.Calls).Should(BeEmpty())
 		})
 	})
 })
