@@ -18,13 +18,14 @@ import (
 // The branch is where the rewrite is active. If the branch doesn't
 // yield a result, the normal resolving is continued.
 type RewriterResolver struct {
+	configurable[*config.RewriterConfig]
 	NextResolver
-	rewrite          map[string]string
-	inner            Resolver
-	fallbackUpstream bool
+	typed
+
+	inner Resolver
 }
 
-func NewRewriterResolver(cfg config.RewriteConfig, inner ChainedResolver) ChainedResolver {
+func NewRewriterResolver(cfg config.RewriterConfig, inner ChainedResolver) ChainedResolver {
 	if len(cfg.Rewrite) == 0 {
 		return inner
 	}
@@ -36,27 +37,22 @@ func NewRewriterResolver(cfg config.RewriteConfig, inner ChainedResolver) Chaine
 	inner.Next(NewNoOpResolver())
 
 	return &RewriterResolver{
-		rewrite:          cfg.Rewrite,
-		inner:            inner,
-		fallbackUpstream: cfg.FallbackUpstream,
+		configurable: withConfig(&cfg),
+		typed:        withType("rewrite"),
+
+		inner: inner,
 	}
 }
 
 func (r *RewriterResolver) Name() string {
-	return fmt.Sprintf("%s w/ %s", Name(r.inner), defaultName(r))
+	return fmt.Sprintf("%s w/ %s", Name(r.inner), r.Type())
 }
 
-// Configuration returns current resolver configuration
-func (r *RewriterResolver) Configuration() (result []string) {
-	result = append(result, "rewrite:")
-	for key, val := range r.rewrite {
-		result = append(result, fmt.Sprintf("  %s = \"%s\"", key, val))
-	}
+// LogConfig implements `config.Configurable`.
+func (r *RewriterResolver) LogConfig(logger *logrus.Entry) {
+	LogResolverConfig(r.inner, logger)
 
-	innerCfg := r.inner.Configuration()
-	result = append(result, innerCfg...)
-
-	return result
+	r.cfg.LogConfig(logger)
 }
 
 // Resolve uses the inner resolver to resolve the rewritten query
@@ -79,7 +75,7 @@ func (r *RewriterResolver) Resolve(request *model.Request) (*model.Response, err
 	request.Req = original
 
 	fallbackCondition := err != nil || (response != NoResponse && response.Res.Answer == nil)
-	if r.fallbackUpstream && fallbackCondition {
+	if r.cfg.FallbackUpstream && fallbackCondition {
 		// Inner resolver had no answer, configuration requests fallback, continue with the normal chain
 		logger.WithField("next_resolver", Name(r.next)).Trace("fallback to next resolver")
 
@@ -130,7 +126,7 @@ func (r *RewriterResolver) rewriteRequest(logger *logrus.Entry, request *dns.Msg
 
 			logger.WithFields(logrus.Fields{
 				"domain":  domainOriginal,
-				"rewrite": rewriteKey + ":" + r.rewrite[rewriteKey],
+				"rewrite": rewriteKey + ":" + r.cfg.Rewrite[rewriteKey],
 			}).Debugf("rewriting %q to %q", domainOriginal, domainRewritten)
 		}
 	}
@@ -139,7 +135,7 @@ func (r *RewriterResolver) rewriteRequest(logger *logrus.Entry, request *dns.Msg
 }
 
 func (r *RewriterResolver) rewriteDomain(domain string) (string, string) {
-	for k, v := range r.rewrite {
+	for k, v := range r.cfg.Rewrite {
 		if strings.HasSuffix(domain, "."+k) {
 			newDomain := strings.TrimSuffix(domain, "."+k) + "." + v
 
