@@ -30,6 +30,8 @@ const (
 
 // UpstreamResolver sends request to external DNS server
 type UpstreamResolver struct {
+	typed
+
 	upstream       config.Upstream
 	upstreamClient upstreamClient
 	bootstrap      *Bootstrap
@@ -51,7 +53,7 @@ type httpUpstreamClient struct {
 }
 
 func createUpstreamClient(cfg config.Upstream) upstreamClient {
-	timeout := time.Duration(config.GetConfig().UpstreamTimeout)
+	timeout := config.GetConfig().UpstreamTimeout.ToDuration()
 
 	tlsConfig := tls.Config{
 		ServerName: cfg.Host,
@@ -209,25 +211,30 @@ func newUpstreamResolverUnchecked(upstream config.Upstream, bootstrap *Bootstrap
 	upstreamClient := createUpstreamClient(upstream)
 
 	return &UpstreamResolver{
+		typed: withType("upstream"),
+
 		upstream:       upstream,
 		upstreamClient: upstreamClient,
 		bootstrap:      bootstrap,
 	}
 }
 
-// Configuration return current resolver configuration
-func (r *UpstreamResolver) Configuration() (result []string) {
-	return []string{r.String()}
+// IsEnabled implements `config.Configurable`.
+func (r *UpstreamResolver) IsEnabled() bool {
+	return true
+}
+
+// LogConfig implements `config.Configurable`.
+func (r *UpstreamResolver) LogConfig(logger *logrus.Entry) {
+	logger.Info(r.upstream)
 }
 
 func (r UpstreamResolver) String() string {
-	return fmt.Sprintf("upstream '%s'", r.upstream.String())
+	return fmt.Sprintf("%s '%s'", r.Type(), r.upstream)
 }
 
 // Resolve calls external resolver
 func (r *UpstreamResolver) Resolve(request *model.Request) (response *model.Response, err error) {
-	logger := log.WithPrefix(request.Log, "upstream_resolver")
-
 	ips, err := r.bootstrap.UpstreamIPs(r)
 	if err != nil {
 		return nil, err
@@ -247,7 +254,7 @@ func (r *UpstreamResolver) Resolve(request *model.Request) (response *model.Resp
 			var err error
 			resp, rtt, err = r.upstreamClient.callExternal(request.Req, upstreamURL, request.Protocol)
 			if err == nil {
-				logger.WithFields(logrus.Fields{
+				r.log().WithFields(logrus.Fields{
 					"answer":           util.AnswerToString(resp.Answer),
 					"return_code":      dns.RcodeToString[resp.Rcode],
 					"upstream":         r.upstream.String(),
@@ -260,7 +267,7 @@ func (r *UpstreamResolver) Resolve(request *model.Request) (response *model.Resp
 				return nil
 			}
 
-			return fmt.Errorf("can't resolve request via upstream server %s: %w", upstreamURL, err)
+			return fmt.Errorf("can't resolve request via upstream server %s (%s): %w", r.upstream, upstreamURL, err)
 		},
 		retry.Attempts(retryAttempts),
 		retry.DelayType(retry.FixedDelay),
@@ -272,7 +279,7 @@ func (r *UpstreamResolver) Resolve(request *model.Request) (response *model.Resp
 			return errors.As(err, &netErr) && netErr.Timeout()
 		}),
 		retry.OnRetry(func(n uint, err error) {
-			logger.WithFields(logrus.Fields{
+			r.log().WithFields(logrus.Fields{
 				"upstream":    r.upstream.String(),
 				"upstream_ip": ip.String(),
 				"question":    util.QuestionToString(request.Req.Question),
@@ -285,5 +292,5 @@ func (r *UpstreamResolver) Resolve(request *model.Request) (response *model.Resp
 		return nil, err
 	}
 
-	return &model.Response{Res: resp, Reason: fmt.Sprintf("RESOLVED (%s)", r.upstream.String())}, nil
+	return &model.Response{Res: resp, Reason: fmt.Sprintf("RESOLVED (%s)", r.upstream)}, nil
 }
