@@ -105,7 +105,7 @@ var _ = BeforeSuite(func() {
 	// create server
 	sut, err = NewServer(&config.Config{
 		CustomDNS: config.CustomDNSConfig{
-			CustomTTL: config.Duration(time.Duration(3600) * time.Second),
+			CustomTTL: config.Duration(3600 * time.Second),
 			Mapping: config.CustomDNSMapping{
 				HostIPs: map[string][]net.IP{
 					"custom.lan": {net.ParseIP("192.168.178.55")},
@@ -143,7 +143,7 @@ var _ = BeforeSuite(func() {
 			BlockType: "zeroIp",
 			BlockTTL:  config.Duration(6 * time.Hour),
 		},
-		Upstream: config.UpstreamConfig{
+		Upstream: config.ParallelBestConfig{
 			ExternalResolvers: map[string][]config.Upstream{"default": {upstreamGoogle}},
 		},
 		ClientLookup: config.ClientLookupConfig{
@@ -158,7 +158,7 @@ var _ = BeforeSuite(func() {
 		},
 		CertFile: certPem.Path,
 		KeyFile:  keyPem.Path,
-		Prometheus: config.PrometheusConfig{
+		Prometheus: config.MetricsConfig{
 			Enable: true,
 			Path:   "/metrics",
 		},
@@ -643,7 +643,7 @@ var _ = Describe("Running DNS server", func() {
 			It("start was called 2 times, start should fail", func() {
 				// create server
 				server, err := NewServer(&config.Config{
-					Upstream: config.UpstreamConfig{
+					Upstream: config.ParallelBestConfig{
 						ExternalResolvers: map[string][]config.Upstream{
 							"default": {config.Upstream{Net: config.NetProtocolTcpUdp, Host: "4.4.4.4", Port: 53}},
 						},
@@ -685,7 +685,7 @@ var _ = Describe("Running DNS server", func() {
 			It("stop was called 2 times, start should fail", func() {
 				// create server
 				server, err := NewServer(&config.Config{
-					Upstream: config.UpstreamConfig{
+					Upstream: config.ParallelBestConfig{
 						ExternalResolvers: map[string][]config.Upstream{
 							"default": {config.Upstream{Net: config.NetProtocolTcpUdp, Host: "4.4.4.4", Port: 53}},
 						},
